==> src/notebooks/1. Exploring the test set.ipynb <==
# Test (accident data)

import os
import pandas as pd

data_path = '../../data/'

df = pd.read_csv(os.path.join(data_path, 'test.csv'))

df.head()

df.shape

'Number of accidents in the test set: {:,}'.format(df.shape[0])
