==> src/notebooks/1. Exploring Vehicles.ipynb <==
# Vehicles

import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

%matplotlib inline
plt.style.use('ggplot')
font = { 'family': 'DejaVu Sans', 'weight': 'bold', 'size': 22 }
plt.rc('font', **font)

from IPython.core.pylabtools import figsize
figsize(14, 7)

data_path = '../../data/'

df = pd.read_csv(os.path.join(data_path, 'vehicles.csv'))

df.head()

df.iloc[0, :]

df.shape

'Number of vehicles: {:,}'.format(df.shape[0])

list(df.columns)

df.dtypes

df.isnull().mean()

'Number of accidents in the accidents related to the vehicles: {:,}'.format(len(df['accident_id'].unique()))

df['accident_id'].value_counts()[:20]

vehicles_per_accident = list(df['accident_id'].value_counts())

np.mean(vehicles_per_accident), np.std(vehicles_per_accident)

df['Vehicle_Reference'].value_counts()

df['Vehicle_Type'].value_counts()

df['Towing_and_Articulation'].value_counts()

df['Vehicle_Manoeuvre'].value_counts()

df['Vehicle_Location-Restricted_Lane'].value_counts()

df['Junction_Location'].value_counts()

df['Skidding_and_Overturning'].value_counts()

df['Hit_Object_in_Carriageway'].value_counts()

df['Vehicle_Leaving_Carriageway'].value_counts()

df['Hit_Object_off_Carriageway'].value_counts()

df['1st_Point_of_Impact'].value_counts()

df['Was_Vehicle_Left_Hand_Drive?'].value_counts()

df['Sex_of_Driver'].value_counts()

df['Age_of_Driver'].value_counts()

df['Engine_Capacity_(CC)'].value_counts()

df['Propulsion_Code'].value_counts()

df['Age_of_Vehicle'].value_counts()

df['Driver_IMD_Decile'].value_counts()

df['Driver_Home_Area_Type'].value_counts()

df['Vehicle_IMD_Decile'].value_counts()
