==> src/notebooks/2. Convert vehicles data to text I.ipynb <==
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

### Read data

data_path = '../../data/'

df_accidents = pd.read_csv(os.path.join(data_path, 'accidents.csv'))

df_accidents.head()

df_accidents.shape

df_accidents.iloc[0,:]

df_accidents.index

'Number of accidents in the training set: {:,}'.format(df_accidents.shape[0])

df_vehicles = pd.read_csv(os.path.join(data_path, 'vehicles.csv'))

df_vehicles.head()

### Merge both tables as a left join with up to N vehicles

vehicles_cols = list(df_vehicles.columns)

vehicles_cols

vehicles_cols.remove('accident_id')

vehicles_cols.remove('Vehicle_Reference')

vehicles_cols

len(vehicles_cols)

df_vehicles.iloc[60178, :]['Vehicle_Reference']

def get_vehicles(row):
    print('Reading row: {}'.format(row['accident_id']))
    for idx in df_vehicles[df_vehicles['accident_id'] == row['accident_id']].index.values:
        vehicle_reference = df_vehicles.iloc[idx, :]['Vehicle_Reference']
        for col_name in vehicles_cols:
            row['{}_{}'.format(col_name, vehicle_reference)] = df_vehicles.iloc[idx, :][col_name]
    return row

df_merged = df_accidents.apply(get_vehicles, axis=1)

df_merged.head()

with open('../../data/accidents_with_vehicles.csv', 'w') as f:
    f.write(df_merged.to_csv(index=False))

!wc -l '../../data/accidents_with_vehicles.csv'
