==> src/notebooks/2. Convert vehicles data to text II.ipynb <==
import csv
import pandas as pd
import os

# 1. Group all vehicles by accident_id in a dict{accident_id:[array with all vehicles]}

vehicles = {}

with open("../../data/vehicles.csv") as f:
    
    lines = f.readlines()

    for line in lines[1:]:
    
        csv_row = line.split(',')
        accident_id = int(csv_row[0])
        vehicles.setdefault(accident_id, [])
        vehicles[accident_id].append(csv_row[1:])

'Number of vehicles: {:,}'.format(len(vehicles))

# 2. Write all vehicles info in a CSV file grouped by accident_id

with open("../../data/text.csv", 'w') as csvfile:
    
    writer = csv.writer(csvfile, delimiter=',', quoting=csv.QUOTE_MINIMAL)
    writer.writerow(['accident_id', 'all_vehicles'])

    for accident_id, info in vehicles.items():
        
        # print(len(info))
        # print(accident_id)
        
        all_vehicles = ''
        for vehicle in info:
            # import pdb; pdb.set_trace()
            all_vehicles += ' ' + ' '.join(vehicle)
        
        writer.writerow( [accident_id, all_vehicles] )

data_path = '../../data/'

df_accidents = pd.read_csv(os.path.join(data_path, 'accidents.csv'))

df_accidents.head()

df_accidents.shape

df_text = pd.read_csv(os.path.join(data_path, 'text.csv'))

df_text.shape

df_text.head()

df_merged = pd.merge(df_accidents, df_text, on='accident_id', how='left')

df_merged.head()

with open('../../data/accidents_with_vehicles.csv', 'w') as f:
    f.write(df_merged.to_csv(index=False))
