==> src/notebooks/2. Convert vehicles data to text III.ipynb <==
### Read data

import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

data_path = '../../data/'

### Read data

df_vehicles = pd.read_csv(os.path.join(data_path, 'vehicles.csv'))

df_vehicles.head()

df_vehicles.dtypes

# df_vehicles.groupby('accident_id')['Vehicle_Type'].apply(','.join)

vehicles_cols = list(df_vehicles.columns)

vehicles_cols

vehicles_cols.remove('accident_id')

vehicles_cols.remove('Vehicle_Reference')

**Note:** for ethical considerations we remove the sex of the driver

vehicles_cols.remove('Sex_of_Driver')

vehicles_cols

len(vehicles_cols)

numeric_cols = [
    'Age_of_Driver',
    'Age_of_Vehicle',
]

category_cols = [
    'Vehicle_Type',
    'Towing_and_Articulation',
    'Vehicle_Manoeuvre',
    'Vehicle_Location-Restricted_Lane',
    'Junction_Location',
    'Skidding_and_Overturning',
    'Hit_Object_in_Carriageway',
    'Vehicle_Leaving_Carriageway',
    'Hit_Object_off_Carriageway',
    '1st_Point_of_Impact',
]

dict_vehicles_cols = {}

dict_vehicles_cols['Vehicle_Reference'] = 'count'

for numeric_col in numeric_cols:
    dict_vehicles_cols[numeric_col] = 'sum'

# for category_col in category_cols:
#     dict_vehicles_cols[category_col] = lambda x: category_col + ': ' + x[category_col].join(', ') if category_col in x else ''

df_grouped = df_vehicles.groupby('accident_id').agg(dict_vehicles_cols)

df_grouped.shape

with open('../../data/grouped_vehicles.csv', 'w') as f:
    f.write(df_grouped.to_csv(index=True))
