==> src/notebooks/5. Split Train and Val sets.ipynb <==
# Split train & val sets

import os
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.model_selection import ShuffleSplit

# SEED
np.random.seed(42)

data_path = '../../data'

df = pd.read_csv(os.path.join(data_path, 'features_all.csv'))

df.head()

df.shape

### Feature names

feature_names = list(df.drop(
    [
        'target',
        # 'all_vehicles',
    ], 1).columns)

len(feature_names)

feature_names[0]

with open('../../data/feature_names.txt', 'w') as f:
    f.write('\n'.join(feature_names))

### X and y

X = np.array(df.drop([
    'target',
    # 'all_vehicles'
], 1))

X.shape

'Number of features: {:,}'.format(X.shape[1])

# X_text = df['all_vehicles']

# X_text.shape

# X_text[0]

y = df['target']

y.shape

### Split

X_train, X_val, y_train, y_val = train_test_split(X, y, test_size=0.20, random_state=42)

len(X_train)

# X_text_train, X_text_val, y_text_train, y_text_val = train_test_split(X_text, y, test_size=0.20, random_state=42)

# len(X_text_train)

### Save

Save the arrays in binary format

np.save(os.path.join(data_path, 'X.npy'), X)
np.save(os.path.join(data_path, 'y.npy'), y)

np.save(os.path.join(data_path, 'X_train.npy'), X_train)
np.save(os.path.join(data_path, 'X_val.npy'), X_val)
np.save(os.path.join(data_path, 'y_train.npy'), y_train)
np.save(os.path.join(data_path, 'y_val.npy'), y_val)

# np.save(os.path.join(data_path, 'X_text_train.npy'), X_text_train)
# np.save(os.path.join(data_path, 'X_text_val.npy'), X_text_val)
# np.save(os.path.join(data_path, 'y_text_train.npy'), y_text_train)
# np.save(os.path.join(data_path, 'y_text_val.npy'), y_text_val)
