==> src/notebooks/6. Modelling with TF-IDF.ipynb <==
import os
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
import pandas as pd
import numpy as np
from sklearn.dummy import DummyClassifier
from sklearn.metrics import classification_report, f1_score
from sklearn.svm import SVC

data_path = '../../data/'

X_text_train = np.load(os.path.join(data_path, 'X_text_train.npy'), allow_pickle=True)
X_text_val = np.load(os.path.join(data_path, 'X_text_val.npy'), allow_pickle=True)
y_text_train = np.load(os.path.join(data_path, 'y_text_train.npy'), allow_pickle=True)
y_text_val = np.load(os.path.join(data_path, 'y_text_val.npy'), allow_pickle=True)

X_text_train[0]

# Count Vectorizer
count_vect = CountVectorizer()
# TFIDF transformer
tfidf_transformer = TfidfTransformer()

# Fit and transform the counts 
train_doc_counts = count_vect.fit_transform(X_text_train)
train_docs_tfidf = tfidf_transformer.fit_transform(train_doc_counts)

train_docs_tfidf.shape

# count_vect.vocabulary_

vocab_len = len(count_vect.vocabulary_)

vocab_len

# Transform test docs
val_docs_counts = count_vect.transform(X_text_val)
val_docs_tfidf = tfidf_transformer.transform(val_docs_counts)

# Train TFIDF to dataframe
# train_docs_tfidf_df = docs_to_dataframe(train_docs_tfidf, vocab_len)

# Test TFIDF dataframe
# test_docs_tfidf_df = docs_to_dataframe(test_docs_tfidf, vocab_len)
