==> src/notebooks/10. Hyper-parameter Tuning.ipynb <==
# Hyper-parameter Tuning

import os
import numpy as np
from sklearn.naive_bayes import GaussianNB
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import classification_report, f1_score

data_path = '../../data/'

X_train = np.load(os.path.join(data_path, 'X_train.npy'), allow_pickle=True)
X_val = np.load(os.path.join(data_path, 'X_val.npy'), allow_pickle=True)
y_train = np.load(os.path.join(data_path, 'y_train.npy'), allow_pickle=True)
y_val = np.load(os.path.join(data_path, 'y_val.npy'), allow_pickle=True)

## Grid Search

parameters = {
    "var_smoothing": (1e-18, 1e-9, 1e-1),
}

# Perform grid search
model = GridSearchCV(
    GaussianNB(priors=None),
    parameters,
    cv=5,
    n_jobs=-1,
)

model.fit(X_train, y_train)

y_pred = model.predict(X_val)

f1_score(y_val, y_pred)

print(classification_report(y_val, y_pred))
