==> src/notebooks/11. Test final model.ipynb <==
# Test final model

import os
import pandas as pd
import numpy as np
import pickle

data_path = '../../data/'

df_test = pd.read_csv(os.path.join(data_path, 'test.csv'))

df_test.head()

df_test.shape

'Number of accidents in the test set: {:,}'.format(df_test.shape[0])

df_vehicles = pd.read_csv(os.path.join(data_path, 'grouped_vehicles.csv'))

df_vehicles.shape

df_vehicles.head()

df_vehicles.columns

df_merged = pd.merge(df_test, df_vehicles, on='accident_id', how='left')

df_merged.head()

df_merged.shape

df_merged['date'] = pd.to_datetime(df_merged['date'])

df_merged['day_of_week'] = df_merged['date'].dt.dayofweek

df_merged['month'] = df_merged['date'].dt.month

df_merged['time'] = pd.to_datetime(df_merged['time'], format='%H:%M')

df_merged['hour'] = df_merged['time'].dt.hour

numeric_columns = [
    'number_of_vehicles',
    'number_of_casualties',
    'speed_limit',
    # vehicles
    # 'Vehicle_Reference',
    'Age_of_Driver',
    'Age_of_Vehicle'
]

category_columns = [
    'road_type',
    '1st_road_class',
    '2nd_road_class',
    'urban_or_rural_area',
    'junction_detail',
    'junction_control',
    'pedestrian_crossing-human_control',
    'pedestrian_crossing-physical_facilities',
    'light_conditions',
    'weather_conditions',
    'road_surface_conditions',
    'special_conditions_at_site',
    'carriageway_hazards',
    # own pre-computed features
    'day_of_week',
    'month',
    'hour',
    # from vehicles
    'Vehicle_Type',
    'Towing_and_Articulation',
    'Vehicle_Manoeuvre',
    'Vehicle_Location-Restricted_Lane',
    'Junction_Location',
    'Skidding_and_Overturning',
    'Hit_Object_in_Carriageway',
    'Vehicle_Leaving_Carriageway',
    'Hit_Object_off_Carriageway',
    '1st_Point_of_Impact',
]

df_features = pd.DataFrame()

accident_ids = df_merged['accident_id']

# Numeric
for col_name in numeric_columns:
    
    print('Number: {}'.format(col_name))
    
    df_features[col_name] = df_merged[col_name]

# Categories
for col_name in category_columns:
    
    print('Category: {}'.format(col_name))
    # keys = set(df[col_name].unique())
    # print(keys)
        
    ## One-hot encoding
        
    # Categories
    df_categories = pd.get_dummies(df_merged[col_name], prefix=col_name)

    # Append to features
    df_features = pd.concat([df_features, df_categories], axis=1)

df_features.head()

df_features.iloc[0, :]

len(list(df_features.columns))

### Load data

with open('../../data/training_columns.txt') as f:
    training_columns = f.read().split(';')

len(training_columns)

training_columns

training_columns.remove('target')

len(training_columns)

df_features = df_features[training_columns]

df_features['junction_control_Not at junction or within 20 metres'] = 0

df_features = df_features.fillna(0)

X_test = np.array(df_features)

### Load model

filename = '../../models/finalized_model.sav'
loaded_model = pickle.load(open(filename, 'rb'))

results = loaded_model.predict(X_test)

df_results = pd.DataFrame()
df_results['accident_id'] = accident_ids
df_results['result'] = results

float(sum(results)) / len(results)

with open('../../models/submission.csv', 'w') as f:
    f.write(df_results.to_csv(index=False))
